implemented the initial LIVEdotIO Model/View system
The Backend index.js sends the LDIModel inside the published eventInfoObj (which is the container for the data we want to publish to the Frontend). This model is grabbed by the client in the Frontend and its contents are used to create/update the LDIView (which for now is a simple div with 3 elements inside) in the page. Care must be taken to ensure that the model and the view agree with each other.